Testing and fixing button
