+++ src/segenie_frontend/src/Views/Dashboard/Buildspace/YellowPortal.jsx
@@ -28,7 +28,7 @@ const YellowPortal = ({setClaim, setColor, provider, canClaim}) => {
         <h3 className="text-xl text-[#f77f00]">Yellow Portal</h3>
         <h3>Claim on:</h3>
         <div className={classnames(styles. portfolio__item_cta)}>
-        {canClaim&&
+        {canClaim &&
           <a onClick={() => handleClick()}  href="#about" className='py-2 px-4 rounded-lg bg-gradient-to-r from-[#a855f7] to-[#3b82f6] hover:from-[#4ade80] hover:to-[#3b82f6] '>Jan. 24th</a>
         }
         </div>
